+++ cmd/cyn/root.go
@@ -4,9 +4,11 @@ import (
 	"fmt"
 	"log"
 	"net"
+	"os"
 	"time"
 
 	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
 	"golang.org/x/sync/errgroup"
 
 	"github.com/evertras/cynomys/pkg/listener"
@@ -16,15 +18,38 @@ import (
 var (
 	listenOnUDPList []string
 	sendUDPToList   []string
+	configFilePath  string
 )
 
 func init() {
 	rootCmd.Flags().StringSliceVarP(&listenOnUDPList, "listen-udp", "u", nil, "An IP:port address to listen on for UDP.  Can be specified multiple times.")
 	rootCmd.Flags().StringSliceVarP(&sendUDPToList, "send-udp", "U", nil, "An IP:port address to send to (UDP).  Can be specified multiple times.")
+	rootCmd.Flags().StringVarP(&configFilePath, "config-file", "c", "", "A file path to load as additional configuration.")
 }
 
 var rootCmd = &cobra.Command{
 	RunE: func(cmd *cobra.Command, args []string) error {
+		if configFilePath != "" {
+			file, err := os.Open(configFilePath)
+
+			if err != nil {
+				return fmt.Errorf("failed to open config file %q: %w", configFilePath, err)
+			}
+
+			viper.SetConfigType("yaml")
+			err = viper.ReadConfig(file)
+
+			if err != nil {
+				return fmt.Errorf("failed to read config %q: %w", configFilePath, err)
+			}
+
+			extraUdpListeners := viper.GetStringSlice("listen-udp")
+			listenOnUDPList = append(listenOnUDPList, extraUdpListeners...)
+
+			extraUdpSenders := viper.GetStringSlice("send-udp")
+			sendUDPToList = append(sendUDPToList, extraUdpSenders...)
+		}
+
 		eg := errgroup.Group{}
 		count := 0
